pyproject.toml
[project]
name = "projeto-oc"
version = "0.1.0"
description = "Controle das Organizações Criminosas atuantes no ERS."
authors = [
    {name = "PWSRS",email = "pablo.weber@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
